requirements.txt
cython==0.29.28
click
jaconv
yt_dlp
ffmpeg-python
tinydb
yt-dlp
pillow
opencv-contrib-python
